+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Changed
 
 ### Added
+[PR # 7](https://github.com/spacetelescope/mast_contributor_tools/pull/7)
+    - Adding unit tests for file name checker
 [PR # 3](https://github.com/spacetelescope/mast_contributor_tools/pull/3)
     - GitHub CI/CD action
     - pre-commit-hook installation